// src/store/actions/actions.enum.ts
export type ActionTypes = UiActions | NoteActions | CategoryActions;

export enum UiActions {
  OPEN_NOTE_DIALOG = 'OPEN_NOTE_DIALOG',
  CLOSE_NOTE_DIALOG = 'CLOSE_NOTE_DIALOG',
  OPEN_CONFIRMATION_DIALOG = 'OPEN_CONFIRMATION_DIALOG',
  CLOSE_CONFIRMATION_DIALOG = 'CLOSE_CONFIRMATION_DIALOG',
  OPEN_SIDEBAR = 'OPEN_SIDEBAR',
  CLOSE_SIDEBAR = 'CLOSE_SIDEBAR',
}

export enum NoteActions {
  GET_NOTES = 'GET_NOTES',
  GET_NOTES_SUCCESS = 'GET_NOTES_SUCCESS',
  GET_NOTES_FAIL = 'GET_NOTES_FAIL',
  CREATE_NOTE = 'CREATE_NOTE',
  CREATE_NOTE_SUCCESS = 'CREATE_NOTE_SUCCESS',
  CREATE_NOTE_FAIL = 'CREATE_NOTE_FAIL',
}

export enum CategoryActions {
  GET_CATEGORIES = 'GET_CATEGORIES',
  GET_CATEGORIES_SUCCESS = 'GET_CATEGORIES_SUCCESS',
  GET_CATEGORIES_FAIL = 'GET_CATEGORIES_FAIL',
  ADD_TEMPORARY_CATEGORY = 'ADD_TEMPORARY_CATEGORY',
  DELETE_TEMPORARY_CATEGORY = 'DELETE_TEMPORARY_CATEGORY',
  CREATE_CATEGORY = 'CREATE_CATEGORY',
  CREATE_CATEGORY_SUCCESS = 'CREATE_CATEGORY_SUCCESS',
  CREATE_CATEGORY_FAIL = 'CREATE_CATEGORY_FAIL',
  SELECT_CATEGORY = 'SELECT_CATEGORY',
  EDIT_CATEGORY = 'EDIT_CATEGORY',
  EDIT_CATEGORY_SUCCESS = 'EDIT_CATEGORY_SUCCESS',
  UPDATE_CATEGORY_SUCCESS = 'UPDATE_CATEGORY_SUCCESS',
  UPDATE_CATEGORIES_STARTED = 'UPDATE_CATEGORIES_STARTED',
  UPDATE_CATEGORIES_FINISHED = 'UPDATE_CATEGORIES_FINISHED',
}